Extract AI extract-memories core
